Update MagazineRack.java
okay, I have updated the open and read file, and it is working correctly. however, I am still working on the open and write file. It is still writing the data to the file in incorrect format. Sorry if its a getting pass a bit overdue, but ima rest up and work on it again later today and upload it by later tonight.